requirements.txt
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.0
flask==2.0.1